rename some stuff, fix tricky nil interface situations
